Started using isolate-vm for eval.
